refactor: creating utils module for testing
