+++ api/tests/test_main.py
@@ -4,26 +4,7 @@ sys.path.insert(0, "../")
 
 from app.main import app
 from fastapi.testclient import TestClient
-
-import random
-
-def generate_cpf():
-    cpf = [random.randint(0, 9) for x in range(9)]
-
-    for _ in range(2):
-        val = sum([(len(cpf) + 1 - i) * v for i, v in enumerate(cpf)]) % 11
-
-        cpf.append(11 - val if val > 1 else 0)
-
-    cpf = [str(x) for x in cpf]
-
-    return ''.join(cpf)
-
-def generate_cns():
-    cns = [random.randint(0, 16) for x in range(9)]
-    cns = [str(x) for x in cns]
-
-    return ''.join(cns)
+from utils import generate_cns, generate_cpf
 
 
 def test_auth():
@@ -46,6 +27,7 @@ def test_auth():
     assert 'access_token' in result_body.keys()
 
     return result_body.get('access_token')
+
 
 def test_patient_creation__minimal():
     token = test_auth()
@@ -73,6 +55,7 @@ def test_patient_creation__minimal():
     assert response.json() != None
     assert 'cpf' in response.json()
     assert response.json()['cpf'] == random_cpf
+
 
 def test_patient_creation__complete():
     token = test_auth()
